Preparing for first beta:    Editing now pretty much works    Added playback mode

+++ Prototyper/LinkImageViewController.m
@@ -12,7 +12,7 @@
 
 @interface LinkImageViewController () <UICollectionViewDataSource, UICollectionViewDelegate>
 {
-    int selectedIndex;
+    NSInteger selectedIndex;
     
     UIBarButtonItem *doneBtn;
 }
